fix: remove some unused packages and file upload download old api
