+++ editor/package.json
@@ -28,14 +28,11 @@
     "flexlayout-react": "^0.5.5",
     "graphql": "^16.3.0",
     "immer": "^9.0.12",
-    "jszip": "^3.6.0",
-    "jszip-utils": "^0.1.0",
     "lodash": "^4.17.21",
     "mui-color": "^2.0.0-beta.2",
     "nanoid": "^3.1.30",
     "on-change": "^4.0.0",
     "pixi.js": "^5.3.3",
-    "pngjs": "^6.0.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-colorful": "^5.5.1",
@@ -52,7 +49,6 @@
     "sweetalert2-react-content": "^4.2.0",
     "three": "^0.137.4",
     "throttle-debounce": "^3.0.1",
-    "use-deep-compare-effect": "^1.8.1",
     "use-immer": "^0.6.0",
     "wavesurfer.js": "^6.0.0",
     "ws": "^7.4.3"
